service test 구현 voucher 할당 기능 fix

+++ src/main/java/org/prgrms/part1/engine/domain/Voucher.java
@@ -12,11 +12,11 @@ public interface Voucher {
 
     VoucherType getVoucherType();
 
-    Long getValue();
+    Integer getValue();
 
     LocalDateTime getCreatedAt();
 
-    void changeValue(long value);
+    void changeValue(Integer value);
 
     void changeOwner(Customer customer);
